Move weapon_info to its own file
in prep for refactoring weapon_xml to weapon_factory
