+++ engine/src/cmd/weapon_factory.cpp
@@ -1,0 +1,6 @@
+#include "weapon_factory.h"
+
+weapon_factory::weapon_factory()
+{
+
+}
