+++ engine/src/cmd/weapon_factory.h
@@ -1,0 +1,11 @@
+#ifndef WEAPON_FACTORY_H
+#define WEAPON_FACTORY_H
+
+
+class weapon_factory
+{
+public:
+    weapon_factory();
+};
+
+#endif // WEAPON_FACTORY_H
